pyproject.toml
[project]
name = "job-boo"
version = "0.1.0"
description = "AI-powered job search, resume tailoring, and application automation"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1,<9",
    "pyyaml>=6.0.2,<7",
    "pymupdf>=1.26.7,<2",
    "httpx>=0.27,<1",
    "anthropic>=0.80,<1",
    "openai>=1.50,<2",
    "beautifulsoup4>=4.12,<5",
    "rich>=13.0,<15",
]

[project.optional-dependencies]
jobspy = ["python-jobspy>=1.1,<2"]
test = ["pytest>=8.0", "pytest-cov>=5.0"]

[project.scripts]
job-boo = "job_boo.cli:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
